+++ beta/layouts/discover/index.js
