+++ beta/routes.js
@@ -16,7 +16,7 @@ module.exports = (app) => {
   app.route('/u/:id/library/history', Layout(require('./views/library/history')))
   app.route('/u/:id/library/:type', Layout(require('./views/library')))
   app.route('/u/:id/membership', Layout(require('./views/profile/membership')))
-  app.route('/discovery', Layout(require('./views/discovery')))
+  app.route('/discover', Layout(require('./views/discover')))
   app.route('/releases', Layout(require('./views/releases')))
   app.route('/tracks', Layout(require('./views/tracks')))
   app.route('/labels', Layout(require('./views/labels')))
