+++ beta/views/main.js
@@ -2,7 +2,7 @@ const html = require('choo/html')
 const viewLayout = require('../layouts/default')
 
 /**
- * This view is currently a placeholder. We only redirects to / or /discovery
+ * This view is currently a placeholder. We only redirects to / or /discover
  */
 
 module.exports = () => viewLayout(renderMain)
